week1

+++ main-3-5.cpp
@@ -2,10 +2,10 @@
 
 // The `extern` keyword tells the compiler that the count function exists,
 // but will be implemented somewhere else
-extern int sumeven(int[], int);
+extern int sumeven(double[], int);
 
 int main() {
-    int array[5] = {4,5,6,7,8};
+    double array[5] = {4,5,6,7,8};
     std::cout << "The number is: " << sumeven(array, 5) << std::endl;
     return 0;
 }
